feat: move AppLayout component to +layout
