+++ src/routes/favourites/+page.svelte
@@ -3,16 +3,15 @@
   import TrashIcon from '@lucide/svelte/icons/trash-2';
   import { listen } from '@tauri-apps/api/event';
   import Gallery from '../../components/Gallery.svelte';
-  import {
-    exportFavorites,
-  } from '../../components/common/ImageOperations';
+  import { exportFavorites } from '../../components/common/ImageOperations';
   import KeyboardShortcuts from '../../components/common/KeyboardShortcuts.svelte';
-  import AppLayout from '../../components/layout/AppLayout.svelte';
   import ConfirmationModal from '../../components/modals/ConfirmationModal.svelte';
   import { favorites, photos } from '../../stores/galleryStore';
-  
+
   // Filter photos to only show favorites
-  const favoritePhotos = $derived($photos.filter(photo => $favorites.has(photo.path)));
+  const favoritePhotos = $derived(
+    $photos.filter((photo) => $favorites.has(photo.path)),
+  );
 
   let exportButtonText = $state('Export Favourites');
   let showDeleteConfirmation = $state(false);
@@ -51,42 +50,42 @@
 
 <KeyboardShortcuts actions={keyboardActions} />
 
-<AppLayout>
-  {#if $favorites.size === 0}
-    <div class="flex flex-col items-center justify-end">
-      <p class="mb-4 text-lg">No favorites</p>
-      <Button href="/">Home</Button>
+{#if $favorites.size === 0}
+  <div class="flex flex-col items-center justify-end">
+    <p class="mb-4 text-lg">No favorites</p>
+    <Button href="/">Home</Button>
+  </div>
+{:else}
+  <!-- Action buttons container -->
+  <div class="sticky top-4 z-10 mb-4 flex justify-center px-4">
+    <div
+      class="flex items-center gap-2 rounded-lg border bg-background/80 p-2 shadow-sm backdrop-blur-sm"
+    >
+      <Button
+        variant="outline"
+        onclick={handleExport}
+        class="text-sm font-medium"
+      >
+        {exportButtonText}
+        <span class="ml-1 rounded-full bg-primary/10 px-2 py-0.5 text-xs">
+          {$favorites.size}
+        </span>
+      </Button>
+      <Button
+        variant="destructive"
+        onclick={showDeleteModal}
+        size="sm"
+        class="px-3"
+      >
+        <TrashIcon class="h-4 w-4" />
+      </Button>
     </div>
-  {:else}
-    <!-- Action buttons container -->
-    <div class="sticky top-4 z-10 mb-4 flex justify-center px-4">
-      <div class="flex items-center gap-2 rounded-lg bg-background/80 p-2 shadow-sm backdrop-blur-sm border">
-        <Button 
-          variant="outline" 
-          onclick={handleExport}
-          class="text-sm font-medium"
-        >
-          {exportButtonText}
-          <span class="ml-1 rounded-full bg-primary/10 px-2 py-0.5 text-xs">
-            {$favorites.size}
-          </span>
-        </Button>
-        <Button 
-          variant="destructive" 
-          onclick={showDeleteModal}
-          size="sm"
-          class="px-3"
-        >
-          <TrashIcon class="h-4 w-4" />
-        </Button>
-      </div>
-    </div>
-    
-    <Gallery photos={favoritePhotos} />
-  {/if}
-</AppLayout>
+  </div>
 
-<ConfirmationModal 
+  <Gallery photos={favoritePhotos} />
+{/if}
+
+<ConfirmationModal
   bind:open={showDeleteConfirmation}
   title="Clear All Favourites"
   description="Are you sure you want to clear all favourites? This action cannot be undone."
@@ -95,5 +94,7 @@
   variant="destructive"
   icon={TrashIcon}
   onConfirm={handleClearFavorites}
-  onCancel={() => { showDeleteConfirmation = false; }}
+  onCancel={() => {
+    showDeleteConfirmation = false;
+  }}
 />
